uint8_t instead of short to speed up a little more

+++ include/LDPC/NBLLR.hpp
@@ -3,11 +3,13 @@
  * @author Yifan Zhang (scirocco_gti@yeah.net)
  * @brief
  * @date 2021-04-20 00:30:58
- * @modified: 2021-04-27 13:09:59
+ * @modified: 2021-04-27 15:18:00
  */
 
 #ifndef NBLLR_HPP
 #define NBLLR_HPP
+
+#include <cstdint>
 
 /**
  * @brief class for NBLLR, has two value: 1. LLR; 2. corresponding Q
@@ -16,12 +18,12 @@
 class NBLLR {
   private:
     float data;
-    int Q;  // should in [1, GF]
+    uint8_t Q;  // should in [1, GF]
   public:
-    NBLLR(float d, int q);
+    NBLLR(float d, uint8_t q);
     ~NBLLR();
     float getLLR() const;
-    int getQ();
+    uint8_t getQ();
 };
 
 bool operator<(const NBLLR& llr1, const NBLLR& llr2);
